Moved group tests to group package
