+++ packages/creator/src/createItemTemplate.ts
@@ -310,6 +310,7 @@ export function createItemTemplate(
               itemId,
               common.createPlaceholderTemplate(itemId, "unknown")
             );
+            resolve(true);
           }
         );
     }
